debug

+++ channel.php
@@ -19,6 +19,8 @@
     }
 
     $subed = FALSE;
+    $friends = FALSE;
+    $pending = FALSE;
     $media= [];
     $query = "SELECT mediaID, mediaType, title, path, description FROM Media WHERE (uploadUser = '$channelID')";
     $results = mysqli_query($mysqli, $query);
@@ -64,13 +66,53 @@
             $userID = $_COOKIE['user'];
             $sub = filter_input(INPUT_POST, 'sub', FILTER_SANITIZE_STRING);
             if($sub) {
-                $query = "INSERT INTO Subscription (subscriber, subscribee) VALUES 
-                ('{$userID}', '{$channelID}')";
-                mysqli_query($mysqli, $query) or die(mysqli_error($mysqli));
+                $query = "SELECT * FROM Subscription WHERE subscriber='$userID' AND subscribee='$channelID'";
+                $result = mysqli_query($mysqli, $query) or die(mysqli_error($mysqli));
+                if($result->num_rows == 0) {
+                    $query = "INSERT INTO Subscription (subscriber, subscribee) VALUES 
+                    ('{$userID}', '{$channelID}')";
+                    mysqli_query($mysqli, $query) or die(mysqli_error($mysqli));
+                }
             } else {
-                $query = "DELETE FROM Subscription WHERE 
-                subscriber = '$userID' AND subscribee = '$channelID'";
-                mysqli_query($mysqli, $query) or die(mysqli_error($mysqli));
+                $query = "SELECT * FROM Subscription WHERE subscriber='$userID' AND subscribee='$channelID'";
+                $result = mysqli_query($mysqli, $query) or die(mysqli_error($mysqli));
+                if($result->num_rows > 0) {
+                    $query = "DELETE FROM Subscription WHERE 
+                    subscriber = '$userID' AND subscribee = '$channelID'";
+                    mysqli_query($mysqli, $query) or die(mysqli_error($mysqli));
+                }
+            }
+
+            $friend = filter_input(INPUT_POST, 'friend', FILTER_SANITIZE_STRING);
+            if($friend) {
+                echo "here";
+                /* $query = "SELECT * FROM Relation WHERE (uname1 = '$channelID' AND uname2 = '$userID' AND status = 1)
+                OR (uname2 = '$channelID' AND uname1 = '$userID' AND status = 1)";
+                $result = mysqli_query($mysqli, $query) or die(mysqli_error($mysqli));
+                if($result->num_rows >= 1)
+                    $pending = TRUE;
+                if($result->num_rows == 0) {
+                    $query = "SELECT * FROM Relation WHERE (uname1 = '$channelID' AND uname2 = '$userID' AND status = 2)
+                    OR (uname2 = '$channelID' AND uname1 = '$userID' AND status = 2)";
+                    $result = mysqli_query($mysqli, $query) or die(mysqli_error($mysqli));
+                    if($result->num_rows == 0) {
+                        $query = "INSERT INTO Relation (uname1, uname2, status) VALUES 
+                        ('{$userID}', '{$channelID}', 1)";
+                        mysqli_query($mysqli, $query) or die(mysqli_error($mysqli));
+                    }
+                } */
+            } else {
+                echo "not friend";
+                /* $query = "SELECT * FROM Relation WHERE (uname1 = '$channelID' AND uname2 = '$userID' AND status = 0)
+                OR (uname2 = '$channelID' AND uname1 = '$userID' AND status = )";
+                $result = mysqli_query($mysqli, $query) or die(mysqli_error($mysqli));
+                if($result->num_rows >= 1)
+                    $pending = TRUE;
+                if($result->num_rows > 0) {
+                    $query = "DELETE FROM Subscription WHERE 
+                    subscriber = '$userID' AND subscribee = '$channelID'";
+                    mysqli_query($mysqli, $query) or die(mysqli_error($mysqli));
+                } */
             }
         } else {
             header("Location: /~wcooley/metube/missingcookie.php", true, 301);
@@ -84,6 +126,18 @@
 
         if($result->num_rows >= 1)
             $subed = TRUE;
+        
+        $query = "SELECT * FROM Relation WHERE (uname1 = '$channelID' AND uname2 = '$userID' AND status = 1) OR (uname2 = '$channelID' AND uname1 = '$userID' AND status = 1)";
+        $result = mysqli_query($mysqli, $query) or die(mysqli_error($mysqli));
+        if($result->num_rows >= 1)
+            $pending = TRUE;
+        
+        $query = "SELECT * FROM Relation WHERE (uname1 = '$channelID' AND uname2 = '$userID' AND status = 2) OR (uname2 = '$channelID' AND uname1 = '$userID' AND status = 2)";
+        $result = mysqli_query($mysqli, $query) or die(mysqli_error($mysqli));
+        if($result->num_rows >= 1)
+            $friend = TRUE;
+        
+        
     }
 
 ?>
@@ -133,7 +187,7 @@
                                 echo $channelID;
                             ?>
                         </h5>
-                        <form class="col s6" method="POST">
+                        <form class="col s3" method="POST">
                             <p>
                                 <label>
                                     <?php
@@ -144,6 +198,22 @@
                                         }
                                     ?>
                                     <span>Subscribed</span>
+                                </label>
+                            </p>
+                            <!--<input type="submit" name="submit" value="Submit"/>-->
+                        </form>
+                        <form class="col s3" method="POST">
+                            <p>
+                                <label>
+                                    <?php
+                                        if ($friends) {
+                                            echo "<input onchange='this.form.submit()' name='friend' type='checkbox' checked='checked'/><span>Friends</span>";
+                                        } elseif($pending) {
+                                            echo "<input onchange='this.form.submit()' name='friend' type='checkbox' disabled='disabled'/><span>Friend Request Pending</span>";
+                                        } else {
+                                            echo "<input onchange='this.form.submit()' name='friend' type='checkbox'/><span>Friends</span>";
+                                        }
+                                    ?>
                                 </label>
                             </p>
                             <!--<input type="submit" name="submit" value="Submit"/>-->
